Move `manifest` pkg to a public location
This will allow others consuming OPA as a library to have easier
access to manage/query the bundle manifests.
