+++ bundle/manifest/manifest.go
